Recommender/requirements.txt
tensorflow
tensorflow-recommender
pandas
numpy